requirements.txt
python-dotenv
web3
requests